fix(cli): unify transaction options (#2172)
* fix(cli): unify transaction options

* fix(cli): separate option modules
