+++ cli/src/cmd/forge/create.rs
@@ -3,15 +3,15 @@ use super::verify;
 use crate::{
     cmd::{forge::build::CoreBuildArgs, utils, RetryArgs},
     compile,
-    opts::{EthereumOpts, WalletType},
-    utils::{get_http_provider, parse_ether_value, parse_u256},
+    opts::{EthereumOpts, TransactionOpts, WalletType},
+    utils::get_http_provider,
 };
 use clap::{Parser, ValueHint};
 use ethers::{
     abi::{Abi, Constructor, Token},
     prelude::{artifacts::BytecodeObject, ContractFactory, Middleware},
     solc::info::ContractInfo,
-    types::{transaction::eip2718::TypedTransaction, Chain, U256},
+    types::{transaction::eip2718::TypedTransaction, Chain},
 };
 use eyre::Context;
 use foundry_config::Config;
@@ -50,67 +50,11 @@ pub struct CreateArgs {
     )]
     constructor_args_path: Option<PathBuf>,
 
-    #[clap(
-        long,
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Send a legacy transaction instead of an EIP1559 transaction.",
-        long_help = r#"Send a legacy transaction instead of an EIP1559 transaction.
-
-This is automatically enabled for common networks without EIP1559."#
-    )]
-    legacy: bool,
-
-    #[clap(
-        long = "gas-price",
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Gas price for legacy transactions, or max fee per gas for EIP1559 transactions.",
-        env = "ETH_GAS_PRICE",
-        parse(try_from_str = parse_ether_value),
-        value_name = "PRICE"
-    )]
-    gas_price: Option<U256>,
-
-    #[clap(
-        long,
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Nonce for the transaction.",
-        parse(try_from_str = parse_u256),
-        value_name = "NONCE"
-    )]
-    nonce: Option<U256>,
-
-    #[clap(
-    long = "gas-limit",
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Gas limit for the transaction.",
-        env = "ETH_GAS_LIMIT",
-        parse(try_from_str = parse_u256),
-        value_name = "GAS_LIMIT"
-    )]
-    gas_limit: Option<U256>,
-
-    #[clap(
-        long = "priority-fee",
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Gas priority fee for EIP1559 transactions.",
-        env = "ETH_GAS_PRIORITY_FEE", parse(try_from_str = parse_ether_value),
-        value_name = "PRICE"
-    )]
-    priority_fee: Option<U256>,
-    #[clap(
-        long,
-        help_heading = "TRANSACTION OPTIONS",
-        help = "Ether to send in the transaction.",
-        long_help = r#"Ether to send in the transaction, either specified in wei, or as a string with a unit type.
-
-Examples: 1ether, 10gwei, 0.01ether"#,
-        parse(try_from_str = parse_ether_value),
-        value_name = "VALUE"
-    )]
-    value: Option<U256>,
-
     #[clap(flatten, next_help_heading = "BUILD OPTIONS")]
     opts: CoreBuildArgs,
+
+    #[clap(flatten, next_help_heading = "TRANSACTION OPTIONS")]
+    tx: TransactionOpts,
 
     #[clap(flatten, next_help_heading = "ETHEREUM OPTIONS")]
     eth: EthereumOpts,
@@ -223,12 +167,12 @@ impl CreateArgs {
                     e
                 }
             })?;
-        let is_legacy =
-            self.legacy || Chain::try_from(chain).map(|x| Chain::is_legacy(&x)).unwrap_or_default();
+        let is_legacy = self.tx.legacy ||
+            Chain::try_from(chain).map(|x| Chain::is_legacy(&x)).unwrap_or_default();
         let mut deployer = if is_legacy { deployer.legacy() } else { deployer };
 
         // set tx value if specified
-        if let Some(value) = self.value {
+        if let Some(value) = self.tx.value {
             deployer.tx.set_value(value);
         }
 
@@ -237,22 +181,22 @@ impl CreateArgs {
         provider.fill_transaction(&mut deployer.tx, None).await?;
 
         // set gas price if specified
-        if let Some(gas_price) = self.gas_price {
+        if let Some(gas_price) = self.tx.gas_price {
             deployer.tx.set_gas_price(gas_price);
         }
 
         // set gas limit if specified
-        if let Some(gas_limit) = self.gas_limit {
+        if let Some(gas_limit) = self.tx.gas_limit {
             deployer.tx.set_gas(gas_limit);
         }
 
         // set nonce if specified
-        if let Some(nonce) = self.nonce {
+        if let Some(nonce) = self.tx.nonce {
             deployer.tx.set_nonce(nonce);
         }
 
         // set priority fee if specified
-        if let Some(priority_fee) = self.priority_fee {
+        if let Some(priority_fee) = self.tx.priority_gas_price {
             if is_legacy {
                 panic!("there is no priority fee for legacy txs");
             }
